example/requirements.txt
tensorflow
pillow
protobuf
